Add Reviews edit

+++ Maets/Controllers/ReviewsController.cs
@@ -35,23 +35,14 @@ namespace Maets.Controllers
         }
 
         // GET: Reviews/Details/5
-        public async Task<IActionResult> Details(Guid? id)
+        public IActionResult Details(Guid? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var review = await _context.Reviews
-                .Include(r => r.App)
-                .Include(r => r.Author)
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (review == null)
-            {
-                return NotFound();
-            }
-
-            return View(_mapper.Map<ReviewReadDto>(review));
+            return RedirectToPage("/store/review", new { Id = id.Value });
         }
 
         // GET: Reviews/Create
@@ -122,61 +113,39 @@ namespace Maets.Controllers
                 return NotFound();
             }
 
-            var review = await _context.Reviews.FindAsync(id);
-            if (review == null)
-            {
-                return NotFound();
-            }
-            ViewData["AppId"] = new SelectList(_context.Apps, "Id", "Id", review.AppId);
-            ViewData["AuthorId"] = new SelectList(_context.Users, "Id", "Id", review.AuthorId);
-            return View(review);
+            var review = await _context.Reviews
+                .Include(x => x.App)
+                .Include(x => x.Author)
+                .FirstOrDefaultAsync(x => x.Id == id.Value);
+            
+            return View(_mapper.Map<ReviewWriteDto>(review));
         }
 
         // POST: Reviews/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         [Authorize]
-        public async Task<IActionResult> Edit(Guid id, [Bind("Score,Title,Description,AuthorId,AppId,Id")] Review review)
+        public async Task<IActionResult> Edit(Guid id, ReviewWriteDto reviewDto)
         {
-            if (id != review.Id)
-            {
-                return NotFound();
-            }
-
             if (!ModelState.IsValid)
             {
-                ViewData["AppId"] = new SelectList(_context.Apps, "Id", "Id", review.AppId);
-                ViewData["AuthorId"] = new SelectList(_context.Users, "Id", "Id", review.AuthorId);
-                return View(review);
+                return View(reviewDto);
             }
             
-            _context.Update(review);
+            var review = await _context.Reviews
+                .Include(x => x.App)
+                .Include(x => x.Author)
+                .FirstOrDefaultAsync(x => x.Id == id);
+
+            if (review is null)
+            {
+                throw new NotFoundException<Review>();
+            }
+
+            _mapper.Map(reviewDto, review);
             await _context.SaveChangesAsync();
             
             return RedirectToAction(nameof(Index));
-        }
-
-        // GET: Reviews/Delete/5
-        [Authorize]
-        public async Task<IActionResult> Delete(Guid? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var review = await _context.Reviews
-                .Include(r => r.App)
-                .Include(r => r.Author)
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (review == null)
-            {
-                return NotFound();
-            }
-
-            return View(review);
         }
 
         // POST: Reviews/Delete/5
@@ -186,8 +155,12 @@ namespace Maets.Controllers
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
             var review = await _context.Reviews.FindAsync(id);
-            _context.Reviews.Remove(review);
-            await _context.SaveChangesAsync();
+            if (review is not null)
+            {
+                _context.Reviews.Remove(review);
+                await _context.SaveChangesAsync();
+            }
+            
             return RedirectToAction(nameof(Index));
         }
     }
